+++ src/oci_attr/mod.rs
@@ -37,6 +37,7 @@ pub mod data_type;
 pub mod handle;
 pub mod mode;
 
+#[allow(clippy::missing_safety_doc)]
 pub unsafe trait OciAttr {
     /// [`SvcCtx`], [`Session`], [`Server`] or [`Stmt`].
     /// Other handle and descriptor types are unsupported.